updated websocket handler tests
